Event Fish separated into new field
