+++ src/main/java/io/github/markassk/fishonmcextras/config/ConfigConstants.java
@@ -7,7 +7,6 @@ public class ConfigConstants {
     public static final String OTHER = "other";
     public static final String COSMETIC = "cosmetic";
     public static final String FUN = "fun";
-    public static final String FISHEVENTS = "fishevents";
 
     public static final boolean DEV = false;
 }
