+++ src/main/java/io/github/markassk/fishonmcextras/config/FishOnMCExtrasConfig.java
@@ -149,10 +149,4 @@ public class FishOnMCExtrasConfig implements ConfigData {
     @ConfigEntry.Gui.Excluded
     public boolean isTimerButtonMenuOpen = true;
     //endregion
-
-    //region Fish Events
-    @ConfigEntry.Category(value = FISHEVENTS)
-    @ConfigEntry.Gui.CollapsibleObject
-    public FishEventsConfig.Fishevents fishevents = new FishEventsConfig.Fishevents();
-    //endregion
 }
