+++ src/main/resources/assets/fishonmcextras/lang/en_us.json
@@ -81,6 +81,22 @@
   "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.dryStreakToggles.showPet": "Show Pets",
   "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.dryStreakToggles.showShard": "Show Shards",
   "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.dryStreakToggles.showLightningBottle": "Show Lightning Bottles",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles": "§f§lFish Event Toggles",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.rarityToggles": "§f§lFish Rarity Toggles",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.rarityToggles.showSpecial": "Show §f\uF092",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.rarityToggles.showSpecial.@Tooltip": "§7Make sure you have Rarities §aenabled §7in Rarity Toggles",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.variantToggles": "§f§lFish Variant Toggles",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.variantToggles.showAlternate": "Show §f\uF098",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.variantToggles.showAlternate.@Tooltip": "§7Make sure you have Variants §aenabled in Variant Toggles",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.variantToggles.showSpooky": "Show §f\uF102",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.variantToggles.showSpooky.@Tooltip": "§7Make sure you have Variants §aenabled in Variant Toggles",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.variantToggles.showFrozen": "Show §f\uF179",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.variantToggles.showFrozen.@Tooltip": "§7Make sure you have Variants §aenabled §7in Variant Toggles",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.dryStreakToggles": "§f§lFish Drystreak Toggles",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.dryStreakToggles.showSpecial": "Show §f\uF092",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.dryStreakToggles.showAlternate": "Show §f\uF098",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.dryStreakToggles.showSpooky": "Show §f\uF102",
+  "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.fisheventToggles.dryStreakToggles.showFrozen": "Show §f\uF179",
   "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.otherToggles": "§f§lOther Toggles",
   "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.otherToggles.showPercentages": "Show percentages",
   "text.autoconfig.fishonmcextras.option.fishTracker.fishTrackerToggles.otherToggles.useNewTitle": "Use the New title system.",
@@ -319,24 +335,5 @@
   "text.autoconfig.fishonmcextras.option.fun.biteBobber.@Tooltip": "§7This will §cdisable§7 the bite notification from the HUD",
   "text.autoconfig.fishonmcextras.option.fun.lightOnBobber": "Make bobber emit a light source",
   "text.autoconfig.fishonmcextras.option.fun.lightLevel": "Light level of the light source on bobber",
-  "text.autoconfig.fishonmcextras.option.fun.useCustomTitleScreen": "Use FoE Custom Title Screen",
-
-  "text.autoconfig.fishonmcextras.option.fishevents": "§f§lFish Event Options",
-  "text.autoconfig.fishonmcextras.option.fishevents.rarityToggles": "§f§lFish Rarity Toggles",
-  "text.autoconfig.fishonmcextras.option.fishevents.rarityToggles.showSpecial": "Show §f\uF092",
-  "text.autoconfig.fishonmcextras.option.fishevents.rarityToggles.showSpecial.@Tooltip": "§7Make sure you have Rarities §aenabled §7in Trackers -> Fish Tracker Options -> HUD Toggles -> Rarity Toggles",
-
-  "text.autoconfig.fishonmcextras.option.fishevents.variantToggles": "§f§lFish Variant Toggles",
-  "text.autoconfig.fishonmcextras.option.fishevents.variantToggles.showAlternate": "Show §f\uF098",
-  "text.autoconfig.fishonmcextras.option.fishevents.variantToggles.showAlternate.@Tooltip": "§7Make sure you have Variants §aenabled in §7Trackers -> Fish Tracker Options -> HUD Toggles -> Variant Toggles",
-  "text.autoconfig.fishonmcextras.option.fishevents.variantToggles.showSpooky": "Show §f\uF102",
-  "text.autoconfig.fishonmcextras.option.fishevents.variantToggles.showSpooky.@Tooltip": "§7Make sure you have Variants §aenabled in §7Trackers -> Fish Tracker Options -> HUD Toggles -> Variant Toggles",
-  "text.autoconfig.fishonmcextras.option.fishevents.variantToggles.showFrozen": "Show §f\uF179",
-  "text.autoconfig.fishonmcextras.option.fishevents.variantToggles.showFrozen.@Tooltip": "§7Make sure you have Variants §aenabled §7in Trackers -> Fish Tracker Options -> HUD Toggles -> Variant Toggles",
-
-  "text.autoconfig.fishonmcextras.option.fishevents.dryStreakToggles": "§f§lFish Drystreak Toggles",
-  "text.autoconfig.fishonmcextras.option.fishevents.dryStreakToggles.showSpecial": "Show §f\uF092",
-  "text.autoconfig.fishonmcextras.option.fishevents.dryStreakToggles.showAlternate": "Show §f\uF098",
-  "text.autoconfig.fishonmcextras.option.fishevents.dryStreakToggles.showSpooky": "Show §f\uF102",
-  "text.autoconfig.fishonmcextras.option.fishevents.dryStreakToggles.showFrozen": "Show §f\uF179"
+  "text.autoconfig.fishonmcextras.option.fun.useCustomTitleScreen": "Use FoE Custom Title Screen"
 }
